Copy file new version
This code version contains two algorithms to copy the configuration file in the working directory. Both work on the Cygwin terminal. However, when it is launched in Clion only the algorithm in lines 163-182 in Hybird.cpp actually copies the configuration file in the working directory.

The error might be due to user error (or Clion?)